+++ tests/test_markdown.py
@@ -197,3 +197,21 @@ async def test_link_in_markdown_table_posts_message_when_clicked():
     async with app.run_test() as pilot:
         await pilot.click(Markdown, offset=(3, 3))
         assert app.messages == ["LinkClicked"]
+
+
+async def test_markdown_quoting():
+    # https://github.com/Textualize/textual/issues/3350
+    links = []
+
+    class MyApp(App):
+        def compose(self) -> ComposeResult:
+            self.md = Markdown(markdown="[tété](tété)")
+            yield self.md
+
+        def on_markdown_link_clicked(self, message: Markdown.LinkClicked):
+            links.append(message.href)
+
+    app = MyApp()
+    async with app.run_test() as pilot:
+        await pilot.click(Markdown, offset=(0, 0))
+    assert links == ["tété"]